+++ CHANGELOG.md
@@ -10,6 +10,8 @@ about writing changes to this log.
 
 ## [Unreleased]
 
+- [PR-368](https://github.com/itk-dev/naevnssekretariatet/pull/368)
+  Added `message-uuid` filter on `tvist1:digital-post-envelope:list` command.
 - [PR-367](https://github.com/itk-dev/naevnssekretariatet/pull/367)
   Preselected party relation type upon edit
 - [PR-363](https://github.com/itk-dev/naevnssekretariatet/pull/363)